packaging
Broke out further packaging, pushed to PyPI, and added some docs pages. Also expanded the readme to match the basic doc intro page.
