+++ docs/index.rst
@@ -9,14 +9,19 @@ heliPypter is a package for rotorcraft performance evaluation. Rotorcraft attrib
 and performance characteristics such as Engine Horsepower, Specific Range, and Fuel consumption are evaluated.
 
 heliPypter has an object oriented philosophy, so different rotorcraft configurations can be built, modified,
-and evaluated quickly, with the same methods. The classess have methods for Hover in and out of Ground Effect
+and evaluated quickly, with the same methods. The classes have methods for Hover in and out of Ground Effect
 (HOGE and HIGE), as well as forward flight.
+
+Under the hood, briefly speaking, the code applies Momentum Theory assuming constant chord ideal twist. 
+For forward flight, Gluert's Model is used.
 
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
    
+   analysis
+   theory
    api
    license
