+++ docs/theory.rts
@@ -1,0 +1,4 @@
+Theoretical Background
+======================
+
+Coming soon! Please be patient, I'm a busy grad student with a big-boy job and can only work on this in my free time.
